simulation/requirements.txt
openai>=1.0.0
together>=0.2.0
python-dotenv>=1.0.0
rich>=13.0.0
